+++ config/connection.js
@@ -1,5 +1,6 @@
 const Sequelize = require('sequelize');
 
+
 require('dotenv').config();
 
 // create connection to our db
@@ -10,5 +11,6 @@ const sequelize = process.env.JAWSDB_URL
       dialect: 'mysql',
       port: 3306
     });
+
 
 module.exports = sequelize;
